image datatype

+++ ddpm.py
@@ -70,7 +70,7 @@ class DDPM(nn.Module):
         
         # Transform back to the range [0, 255]
         x = (x.clamp(-1, 1) + 1) / 2  # Scale to [0, 1]
-        x0 = (x * 255).to(torch.uint8)  # Scale to [0, 255] 
+        x0 = (x * 255)  # Scale to [0, 255] 
 
         return x0
